Split FacilitatorHotelView into components
